discovery/multi: move CreateChanAnnouncement to netann
